# docs/ai/context/openapi.yaml
openapi: 3.1.0
info:
  title: Ardtire Governance API
  version: 0.1.0
  description: >
    Canonical governance API for the Ardtire digital governance platform.
    All domain mutations are performed through this service.
servers:
  - url: http://localhost:3002
...
paths:
...
  /gazette/issues/{issueId}/entries:
    post:
      tags: [Gazette]
      summary: Add entry to issue
      operationId: createGazetteEntry
      parameters:
        - $ref: '#/components/parameters/issueId'
      responses:
        '201':
          description: Entry added
  /gazette/issues/{issueId}/actions/publish:
    post:
      tags: [Gazette]
      summary: Publish issue
      operationId: publishIssue
      parameters:
        - $ref: '#/components/parameters/issueId'
      responses:
        '200':
          description: Published
components:
  parameters:
...
    issueId:
      name: issueId
      in: path
      required: true
      schema:
        type: string
...